Remove FixedPointIterator and move back methods
