+++ goalie/mesh_seq.py
@@ -11,7 +11,7 @@ from animate.utility import Mesh
 from firedrake.petsc import PETSc
 from firedrake.pyplot import triplot
 
-from .log import DEBUG, debug, info, logger, warning
+from .log import DEBUG, debug, info, logger, pyrint, warning
 
 __all__ = ["MeshSeq"]
 
@@ -130,7 +130,7 @@ class MeshSeq:
         """
         # TODO #122: Refactor to use the set method
         if not isinstance(meshes, Iterable):
-            meshes = [Mesh(meshes) for subinterval in self.subintervals]
+            meshes = [Mesh(meshes) for subinterval in self.subintervals]  # FIXME
         self.meshes = meshes
         dim = np.array([mesh.topological_dimension() for mesh in meshes])
         if dim.min() != dim.max():
@@ -196,3 +196,46 @@ class MeshSeq:
         if len(axes) == 1:
             axes = axes[0]
         return fig, axes
+
+    def check_element_count_convergence(self):
+        r"""
+        Check for convergence of the fixed point iteration due to the relative
+        difference in element count being smaller than the specified tolerance.
+
+        :return: an array, whose entries are ``True`` if convergence is detected on the
+            corresponding subinterval
+        :rtype: :class:`list` of :class:`bool`\s
+        """
+        if self.params.drop_out_converged:
+            converged = self.converged
+        else:
+            converged = np.array([False] * len(self), dtype=bool)
+        if len(self.element_counts) >= max(2, self.params.miniter + 1):
+            for i, (ne_, ne) in enumerate(zip(*self.element_counts[-2:])):
+                if not self.check_convergence[i]:
+                    self.info(
+                        f"Skipping element count convergence check on subinterval {i})"
+                        f" because check_convergence[{i}] == False."
+                    )
+                    continue
+                if abs(ne - ne_) <= self.params.element_rtol * ne_:
+                    converged[i] = True
+                    if len(self) == 1:
+                        pyrint(
+                            f"Element count converged after {self.fp_iteration+1}"
+                            " iterations under relative tolerance"
+                            f" {self.params.element_rtol}."
+                        )
+                    else:
+                        pyrint(
+                            f"Element count converged on subinterval {i} after"
+                            f" {self.fp_iteration+1} iterations under relative"
+                            f" tolerance {self.params.element_rtol}."
+                        )
+
+        # Check only early subintervals are marked as converged
+        if self.params.drop_out_converged and not converged.all():
+            first_not_converged = converged.argsort()[0]
+            converged[first_not_converged:] = False
+
+        return converged
